+++ SnazzCraft/src/snazzcraft-engine/world/collision.cpp
@@ -8,9 +8,9 @@ SnazzCraft::Voxel* SnazzCraft::World::GetCollidingVoxel(const glm::vec3& Positio
 
     for (int32_t X = ChunkX - 1; X <= ChunkX + 1; X++) {
     for (int32_t Z = ChunkZ - 1; Z <= ChunkZ + 1; Z++) {
-        if (X < 0 || Z < 0 || X >= static_cast<int32_t>(this->Size) || Z >= static_cast<int32_t>(this->Size)) continue;
+        if (X < 0 || Z < 0 || X >= this->Size || Z >= this->Size) continue;
         
-        auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D<int32_t>(X, Z, static_cast<int32_t>(this->Size)));
+        auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D<int32_t>(X, Z, this->Size));
         if (ChunkIterator == this->Chunks.end()) continue;
 
         SnazzCraft::Voxel* CollisionVoxel = ChunkIterator->second->GetCollidingVoxel(Position, Hitbox);
@@ -26,9 +26,9 @@ SnazzCraft::Voxel* SnazzCraft::World::GetCollidingVoxel(const glm::vec3& Positio
     const int32_t ChunkX = static_cast<int32_t>(Position.x / (SnazzCraft::Chunk::Width * SnazzCraft::Voxel::Size));
     const int32_t ChunkZ = static_cast<int32_t>(Position.z / (SnazzCraft::Chunk::Depth * SnazzCraft::Voxel::Size));
 
-    if (ChunkX < 0 || ChunkZ < 0 || ChunkX >= static_cast<int32_t>(this->Size) || ChunkZ >= static_cast<int32_t>(this->Size)) return nullptr;
+    if (ChunkX < 0 || ChunkZ < 0 || ChunkX >= this->Size || ChunkZ >= this->Size) return nullptr;
 
-    auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D<int32_t>(ChunkX, ChunkZ, static_cast<int32_t>(this->Size)));
+    auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D<int32_t>(ChunkX, ChunkZ, this->Size));
     if (ChunkIterator == this->Chunks.end()) return nullptr;
 
     SnazzCraft::Voxel* CollisionVoxel = ChunkIterator->second->GetCollidingVoxel(Position);
