+++ SnazzCraft/src/snazzcraft-engine/world/lighting.cpp
@@ -37,7 +37,7 @@ void SnazzCraft::World::UpdateChunkLighting(SnazzCraft::Chunk* Chunk, bool* Upda
         this->UpdateChunkVerticeLightingAndMesh(I);
     }
 
-    auto ChunkIterator = ChunksToUpdate.find(SnazzCraft::Index2D<int32_t>(Chunk->Position[0], Chunk->Position[1], static_cast<int32_t>(this->Size)));
+    auto ChunkIterator = ChunksToUpdate.find(SnazzCraft::Index2D<int32_t>(Chunk->Position[0], Chunk->Position[1], this->Size));
     if (UpdatedInputChunk != nullptr) *UpdatedInputChunk = ChunkIterator != ChunksToUpdate.end();
 }
 
@@ -49,7 +49,7 @@ void SnazzCraft::World::ApplySunLightingToChunk(SnazzCraft::Chunk* Chunk, std::u
     }
     }
 
-    if (ChunksToUpdate != nullptr) ChunksToUpdate->insert(SnazzCraft::Index2D<int32_t>(Chunk->Position[0], Chunk->Position[1], static_cast<int32_t>(this->Size)));
+    if (ChunksToUpdate != nullptr) ChunksToUpdate->insert(SnazzCraft::Index2D<int32_t>(Chunk->Position[0], Chunk->Position[1], this->Size));
 }
 
 void SnazzCraft::World::ApplySunLightingToColumn(SnazzCraft::Chunk* Chunk, uint32_t LocalChunkX, uint32_t LocalChunkZ, uint32_t StartY, int32_t StartLightValue, std::unordered_set<uint32_t>* ChunksToUpdate)
@@ -79,7 +79,7 @@ void SnazzCraft::World::ApplyLightingVoxel(int32_t LightOrigin[3], int32_t Light
 {
     auto IsOutsideWorld = [this](int32_t X, int32_t Y, int32_t Z) -> bool
     {
-        return X < 0 || Y < 0 || Z < 0 || X >= static_cast<int32_t>(this->Size) * SnazzCraft::Chunk::Width || Y >= SnazzCraft::Chunk::Height || Z >= static_cast<int32_t>(this->Size) * SnazzCraft::Chunk::Depth;
+        return X < 0 || Y < 0 || Z < 0 || X >= this->Size * SnazzCraft::Chunk::Width || Y >= SnazzCraft::Chunk::Height || Z >= this->Size * SnazzCraft::Chunk::Depth;
     };
 
     auto AddLightNodes = [this](std::queue<SnazzCraft::World::LightNode>& Queue, const SnazzCraft::World::LightNode& OriginNode, int LightPropagationDecrease) -> void
@@ -116,11 +116,11 @@ void SnazzCraft::World::ApplyLightingVoxel(int32_t LightOrigin[3], int32_t Light
 
         int32_t ChunkCoordinates[2];
         SnazzCraft::Chunk::GetChunkPosition(CurrentNode.X, CurrentNode.Z, ChunkCoordinates);
-        auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D(ChunkCoordinates[0], ChunkCoordinates[1], static_cast<int32_t>(this->Size)));
+        auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D(ChunkCoordinates[0], ChunkCoordinates[1], this->Size));
         
         if (ChunkIterator == this->Chunks.end()) {
             this->GenerateChunk(ChunkCoordinates[0], ChunkCoordinates[1], true);
-            ChunkIterator = this->Chunks.find(SnazzCraft::Index2D(ChunkCoordinates[0], ChunkCoordinates[1], static_cast<int32_t>(this->Size)));
+            ChunkIterator = this->Chunks.find(SnazzCraft::Index2D(ChunkCoordinates[0], ChunkCoordinates[1], this->Size));
         }
         if (ChunkIterator == this->Chunks.end()) continue;
 
@@ -133,7 +133,7 @@ void SnazzCraft::World::ApplyLightingVoxel(int32_t LightOrigin[3], int32_t Light
 
         if (CurrentNode.LightValue > CurrentExistingValue) {
             ChunkIterator->second->LightValues.insert_or_assign(LightIndex, CurrentNode.LightValue);
-            if (ChunksToUpdate != nullptr) ChunksToUpdate->insert(SnazzCraft::Index2D(ChunkCoordinates[0], ChunkCoordinates[1], static_cast<int32_t>(this->Size)));
+            if (ChunksToUpdate != nullptr) ChunksToUpdate->insert(SnazzCraft::Index2D(ChunkCoordinates[0], ChunkCoordinates[1], this->Size));
             
             int32_t LocalVoxelPosition[3];
             SnazzCraft::Chunk::GetLocalVoxelPosition(CurrentNode.X, CurrentNode.Y, CurrentNode.Z, LocalVoxelPosition);
