+++ SnazzCraft/src/snazzcraft-engine/world/rendering.cpp
@@ -52,9 +52,9 @@ void SnazzCraft::World::RenderChunks() const
 
     for (int32_t X = PlayerChunkPosition[0] - static_cast<int32_t>(this->RenderDistance); X <= PlayerChunkPosition[0] + static_cast<int32_t>(this->RenderDistance); X++) {
     for (int32_t Z = PlayerChunkPosition[1] - static_cast<int32_t>(this->RenderDistance); Z <= PlayerChunkPosition[1] + static_cast<int32_t>(this->RenderDistance); Z++) {
-        if (X < 0 || X >= static_cast<int32_t>(this->Size) || Z < 0 || Z >= static_cast<int32_t>(this->Size)) continue;
+        if (X < 0 || X >= this->Size || Z < 0 || Z >= this->Size) continue;
 
-        auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D<int32_t>(X, Z, static_cast<int32_t>(this->Size)));
+        auto ChunkIterator = this->Chunks.find(SnazzCraft::Index2D<int32_t>(X, Z, this->Size));
         if (ChunkIterator == this->Chunks.end()) continue;
 
         if (ChunkIterator->second->ChunkMesh == nullptr) continue;
